+++ packages/layout/src/tree-layout/data-adapters/html-data.js
@@ -1,11 +1,6 @@
-/* eslint no-useless-constructor: "off" */
 import { DataParser } from './data-parser';
 
 export class HTMLDataAdapter extends DataParser {
-    constructor (data) {
-        super(data);
-    }
-
     getCoordinates () {
         return super.defaultDataPointLogic();
     }
